fix SubstituteVariables problem, and extend IIFEIdioms
Coffeescript generates default constructors which close over _ref2 (which is idiotic) instead of using a local variable.
This was triggering a bug in other closures at the same level.  If another closure didn't have nested functions of its own
that closed over *its* bindings, and it also had a _ref2, it would also use the closed-over _ref2, even if it defined it
itself.  This was leading to incorrect code being generated and the _ref2 object changing out from under a method.

Also extend IIFEIdioms to handle this case:

   (function(){...}).call(this)

Given that we've previously converted all bindings in {...} to let bindings and don't have to worry about hoisting polluting
our enclosing scope, we can replace this with:

   {...}