add --frozen-global flag (see toshok/echo-js#10)
when --frozen-global is present on the command line, it becomes a syntax error to
set a property on the global object (i.e. every lhs of an assignment has to be an
in-scope variable declaration.)

Also, since we know the global object is frozen we can inline loads from the C globals
instead of calling _ejs_global_getprop at all.