add debug info generation to the compiler.
This change adds a -g flag to the ejs script, which turns out debug mode and turns off optimizations
(we don't even run opt in that case.)

The debuginfo metadata shows up in the .ll file but for some reason it's failing verification and
being stripped out by llvm-as.  It also causes some test compilations to fail, so it's definitely
not ready for prime time.

But soon, working stack traces, nice function symbol names, and breakpoints in gdb/lldb.