move the node compat apis from libecho to a separate module.
This does require us to add a --module flag when we're using @node-compat (for now), but it also means that
tests/programs that don't use the apis don't require the module, which is nice.

The following code:

  let os = require('os');
  let path = require('path');
  let fs = require('fs');
  let spawn = require('child_process').spawn;

can now be es6-ified and written as:

  import * as os from            '@node-compat/os';
  import * as path from          '@node-compat/path';
  import * as fs from            '@node-compat/fs';
  import * as child_process from '@node-compat/child_process';
  let spawn = child_process.spawn;

The last 2 lines would become 'import { spawn } ...' when native modules can explicitly list their exports.