we can't rely on coffeescript being installed on ubuntu built hosts.  sigh.
trusty ships with coffeescript 1.4, and we can't install a newer version in
debian/rules, so we have two options:

1) include coffeescript and all its deps in our pkg-src
2) include the compiled .js files in lib/generated in our pkg-src.

going with #2 since it's less code, and the coffeescript version is not
meant to be used for anything but self hosting anyway.  If people need
to make changes to the actual CS they should be working from git instead
of tarballs.  This lets us remove the Build-Depends on coffeescript entirely.

Since we aren't depending on coffeescript, the toplevel ejs file gets moved
to lib/ejs.coffee and gets compiled using the normal CS pipeline.  The toplevel
ejs then becomes a script that execs node lib/generate/ejs.js.  had to also
fix up some --srcdir paths in the CS given the new location of ejs.js.

good times.