permit null's to exist post-node visitor pass over an array.
unfortunately going to have to revisit the behavior of the node visitor and subclasses.
Right now visitArray checks if the return value is null, and if it is it removes the
null from the resulting array.  This breaks the behavior of array holes in ArrayPatterns,
though (they're represented as 'null'.)  For the time being, add another function,
visitArrayKeep, which keeps nulls around, and use that for array pattern elements.  Fixes
kangax's 'destructuring with arrays' test.